First commit to radar SLAM repository, still in development
This repo currently serves as a sandbox for a radar SLAM implementation
using data from the self-driving testbed.  This data was recorded using
rosbag and is assumed to exist in a bagfiles directory in the user's
home directory. The bagfile is used for development, but the SLAM can
be run online when ready. For now, this repo is specific to the car case,
but its purpose may evolve and become more general over time.