Add filtering of targets based on relative speed, tune params
Filtering out targets which have greater than a specified speed relative to the car.
This seems to help a bit espcially with noisy targets around corners, but nothing
major.  Attempted to tune other parameters as well.  It may come down to collecting
better data with T79 or waiting for K79, unfortunately.